Tidied tests and broke graphql out - added new graph func and tests
