+++ python/python/raphtory/__init__.pyi
@@ -8,8 +8,10 @@
 ###############################################################################
 
 class AlgorithmResult:
+
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def get(self, key):
         """
         Returns the value corresponding to the provided key
@@ -17,6 +19,7 @@ class AlgorithmResult:
         Arguments:
             key: The key of type `H` for which the value is to be retrieved.
         """
+
     def get_all(self):
         """
         Returns a Dict containing all the nodes (as keys) and their corresponding values (values) or none.
@@ -24,8 +27,10 @@ class AlgorithmResult:
         Returns:
             A dict of nodes and their values
         """
+
     def get_all_values(self):
         """Returns a a list of all values"""
+
     def get_all_with_names(self):
         """
         Returns a dict with node names and values
@@ -33,6 +38,7 @@ class AlgorithmResult:
         Returns:
             a dict with node names and values
         """
+
     def group_by(self):
         """
         Groups the `AlgorithmResult` by its values.
@@ -41,12 +47,16 @@ class AlgorithmResult:
             A `HashMap` where keys are unique values from the `AlgorithmResult` and values are vectors
             containing keys of type `H` that share the same value.
         """
+
     def max(self):
         """Returns a tuple of the max result with its key"""
+
     def median(self):
         """Returns a tuple of the median result with its key"""
+
     def min(self):
         """Returns a tuple of the min result with its key"""
+
     def sort_by_node(self, reverse=True):
         """
         Sorts by node id in ascending or descending order.
@@ -57,6 +67,7 @@ class AlgorithmResult:
         Returns:
             A sorted list of tuples containing node names and values.
         """
+
     def sort_by_node_name(self, reverse=True):
         """
         The function `sort_by_node_name` sorts a vector of tuples containing a node and an optional
@@ -70,6 +81,7 @@ class AlgorithmResult:
         Returns:
             The function sort_by_node_name returns a vector of tuples. Each tuple contains a Node and value
         """
+
     def sort_by_value(self, reverse=True):
         """
         Sorts the `AlgorithmResult` by its values in ascending or descending order.
@@ -80,6 +92,7 @@ class AlgorithmResult:
         Returns:
             A sorted vector of tuples containing keys of type `H` and values of type `Y`.
         """
+
     def to_df(self):
         """
         Creates a dataframe from the result
@@ -87,8 +100,10 @@ class AlgorithmResult:
         Returns:
             A `pandas.DataFrame` containing the result
         """
+
     def to_string(self):
         """Returns a formatted string representation of the algorithm."""
+
     def top_k(self, k, percentage=False, reverse=True):
         """
         Retrieves the top-k elements from the `AlgorithmResult` based on its values.
@@ -110,12 +125,14 @@ class ConstProperties:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def as_dict(self):
         """
         as_dict() -> dict[str, Any]
 
         convert the properties view to a python dict
         """
+
     def get(self, key):
         """
         get(key: str) -> Any | None
@@ -125,18 +142,21 @@ class ConstProperties:
 
         get property value by key (returns `None` if key does not exist)
         """
+
     def items(self):
         """
         items() -> list[tuple[str, Any]]
 
         lists the property keys together with the corresponding value
         """
+
     def keys(self):
         """
         keys() -> list[str]
 
         lists the available property keys
         """
+
     def values(self):
         """
         values() -> list[Any]
@@ -145,32 +165,36 @@ class ConstProperties:
         """
 
 class DiskGraphStorage:
+
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
-    def graph_dir(self): ...
+
+    def graph_dir(self):
+        ...
+
     @staticmethod
-    def load_from_dir(graph_dir): ...
+    def load_from_dir(graph_dir):
+        ...
+
     @staticmethod
-    def load_from_pandas(graph_dir, edge_df, time_col, src_col, dst_col): ...
+    def load_from_pandas(graph_dir, edge_df, time_col, src_col, dst_col):
+        ...
+
     @staticmethod
-    def load_from_parquets(
-        graph_dir,
-        layer_parquet_cols,
-        node_properties,
-        chunk_size,
-        t_props_chunk_size,
-        read_chunk_size,
-        concurrent_files,
-        num_threads,
-        node_type_col,
-    ): ...
+    def load_from_parquets(graph_dir, layer_parquet_cols, node_properties, chunk_size, t_props_chunk_size, read_chunk_size, concurrent_files, num_threads, node_type_col):
+        ...
+
     def merge_by_sorted_gids(self, other, graph_dir):
         """
         Merge this graph with another `DiskGraph`. Note that both graphs should have nodes that are
         sorted by their global ids or the resulting graph will be nonsense!
         """
-    def to_events(self): ...
-    def to_persistent(self): ...
+
+    def to_events(self):
+        ...
+
+    def to_persistent(self):
+        ...
 
 class Edge:
     """
@@ -180,6 +204,7 @@ class Edge:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def after(self, start):
         """
          Create a view of the Edge including all events after `start` (exclusive).
@@ -190,6 +215,7 @@ class Edge:
         Returns:
              A Edge object.
         """
+
     def at(self, time):
         """
          Create a view of the Edge including all events at `time`.
@@ -200,6 +226,7 @@ class Edge:
         Returns:
              A Edge object.
         """
+
     def before(self, end):
         """
          Create a view of the Edge including all events before `end` (exclusive).
@@ -210,6 +237,7 @@ class Edge:
         Returns:
              A Edge object.
         """
+
     @property
     def date_time(self):
         """
@@ -218,12 +246,14 @@ class Edge:
         Returns:
             (datetime) the datetime of an exploded edge
         """
+
     def default_layer(self):
         """
          Return a view of Edge containing only the default edge layer
         Returns:
              Edge: The layered view
         """
+
     def deletions(self):
         """
         Returns a list of timestamps of when an edge is deleted
@@ -231,6 +261,7 @@ class Edge:
         Returns:
             A list of unix timestamps
         """
+
     def deletions_data_time(self):
         """
         Returns a list of timestamps of when an edge is deleted
@@ -238,9 +269,11 @@ class Edge:
         Returns:
             A list of DateTime objects
         """
+
     @property
     def dst(self):
         """Returns the destination node of the edge."""
+
     @property
     def earliest_date_time(self):
         """
@@ -249,6 +282,7 @@ class Edge:
         Returns:
             the earliest datetime of an edge
         """
+
     @property
     def earliest_time(self):
         """
@@ -257,6 +291,7 @@ class Edge:
         Returns:
             (int) The earliest time of an edge
         """
+
     @property
     def end(self):
         """
@@ -265,6 +300,7 @@ class Edge:
         Returns:
            The latest time that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -273,6 +309,7 @@ class Edge:
         Returns:
              The latest datetime that this Edge is valid or None if the Edge is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Edge containing all layers except the excluded `name`
@@ -284,6 +321,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Edge containing all layers except the excluded `names`
@@ -295,6 +333,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Edge containing all layers except the excluded `name`
@@ -304,6 +343,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Edge containing all layers except the excluded `names`
@@ -313,6 +353,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -325,11 +366,16 @@ class Edge:
         Returns:
             A `WindowSet` object.
         """
+
     def explode(self):
         """Explodes an edge and returns all instances it had been updated as seperate edges"""
-    def explode_layers(self): ...
+
+    def explode_layers(self):
+        ...
+
     def has_layer(self, name):
-        """Check if Edge has the layer `"name"`"""
+        """ Check if Edge has the layer `"name"`"""
+
     def history(self):
         """
         Returns a list of timestamps of when an edge is added or change to an edge is made.
@@ -338,6 +384,7 @@ class Edge:
             A list of unix timestamps.
 
         """
+
     def history_date_time(self):
         """
         Returns a list of timestamps of when an edge is added or change to an edge is made.
@@ -346,15 +393,20 @@ class Edge:
             A list of timestamps.
 
         """
+
     @property
     def id(self):
         """The id of the edge."""
+
     def is_deleted(self):
         """Check if the edge is currently deleted"""
+
     def is_self_loop(self):
         """Check if the edge is on the same node"""
+
     def is_valid(self):
         """Check if the edge is currently valid (i.e., not deleted)"""
+
     @property
     def latest_date_time(self):
         """
@@ -363,6 +415,7 @@ class Edge:
         Returns:
             (datetime) the latest datetime of an edge
         """
+
     @property
     def latest_time(self):
         """
@@ -371,6 +424,7 @@ class Edge:
         Returns:
             (int) The latest time of an edge
         """
+
     def layer(self, name):
         """
          Return a view of Edge containing the layer `"name"`
@@ -379,6 +433,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     @property
     def layer_name(self):
         """
@@ -387,6 +442,7 @@ class Edge:
         Returns:
             (List<str>) The name of the layer
         """
+
     @property
     def layer_names(self):
         """
@@ -395,6 +451,7 @@ class Edge:
         Returns:
             (List<str>) The name of the layer
         """
+
     def layers(self, names):
         """
          Return a view of Edge containing all layers `names`
@@ -406,9 +463,11 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     @property
     def nbr(self):
         """Returns the node at the other end of the edge (same as `dst()` for out-edges and `src()` for in-edges)"""
+
     @property
     def properties(self):
         """
@@ -417,6 +476,7 @@ class Edge:
         Returns:
           Properties on the Edge.
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -430,6 +490,7 @@ class Edge:
         Returns:
             A `WindowSet` object.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -439,6 +500,7 @@ class Edge:
         Returns:
              A Edge object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -449,6 +511,7 @@ class Edge:
         Returns:
              A Edge object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -456,9 +519,11 @@ class Edge:
         Arguments:
 
         """
+
     @property
     def src(self):
         """Returns the source node of the edge."""
+
     @property
     def start(self):
         """
@@ -467,6 +532,7 @@ class Edge:
         Returns:
             The earliest time that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -475,6 +541,7 @@ class Edge:
         Returns:
              The earliest datetime that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def time(self):
         """
@@ -483,6 +550,7 @@ class Edge:
         Returns:
             (int) The time of an exploded edge
         """
+
     def valid_layers(self, names):
         """
          Return a view of Edge containing all layers `names`
@@ -494,6 +562,7 @@ class Edge:
         Returns:
              Edge: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Edge including all events between `start` (inclusive) and `end` (exclusive)
@@ -505,15 +574,17 @@ class Edge:
         Returns:
         r    A Edge object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Edge"""
+        """ Get the window size (difference between start and end) for this Edge"""
 
 class Edges:
     """A list of edges that can be iterated over."""
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def after(self, start):
         """
          Create a view of the Edges including all events after `start` (exclusive).
@@ -524,6 +595,7 @@ class Edges:
         Returns:
              A Edges object.
         """
+
     def at(self, time):
         """
          Create a view of the Edges including all events at `time`.
@@ -534,6 +606,7 @@ class Edges:
         Returns:
              A Edges object.
         """
+
     def before(self, end):
         """
          Create a view of the Edges including all events before `end` (exclusive).
@@ -544,6 +617,7 @@ class Edges:
         Returns:
              A Edges object.
         """
+
     def collect(self):
         """
          Collect all edges into a list
@@ -551,8 +625,10 @@ class Edges:
         Returns:
              list[Edge]: the list of edges
         """
+
     def count(self):
         """Returns the number of edges"""
+
     @property
     def date_time(self):
         """
@@ -561,12 +637,14 @@ class Edges:
         Returns:
            A list of date times.
         """
+
     def default_layer(self):
         """
          Return a view of Edges containing only the default edge layer
         Returns:
              Edges: The layered view
         """
+
     def deletions(self):
         """
         Returns all timestamps of edges where an edge is deleted
@@ -574,6 +652,7 @@ class Edges:
         Returns:
             A list of lists of unix timestamps
         """
+
     def deletions_date_time(self):
         """
         Returns all timestamps of edges where an edge is deleted
@@ -581,9 +660,11 @@ class Edges:
         Returns:
             A list of lists of DateTime objects
         """
+
     @property
     def dst(self):
         """Returns the destination node of the edge."""
+
     @property
     def earliest_date_time(self):
         """
@@ -592,6 +673,7 @@ class Edges:
         Returns:
          Earliest date time of the edges.
         """
+
     @property
     def earliest_time(self):
         """
@@ -600,6 +682,7 @@ class Edges:
         Returns:
         Earliest time of the edges.
         """
+
     @property
     def end(self):
         """
@@ -608,6 +691,7 @@ class Edges:
         Returns:
            The latest time that this Edges is valid or None if the Edges is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -616,6 +700,7 @@ class Edges:
         Returns:
              The latest datetime that this Edges is valid or None if the Edges is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Edges containing all layers except the excluded `name`
@@ -627,6 +712,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Edges containing all layers except the excluded `names`
@@ -638,6 +724,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Edges containing all layers except the excluded `name`
@@ -647,6 +734,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Edges containing all layers except the excluded `names`
@@ -656,6 +744,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -668,11 +757,16 @@ class Edges:
         Returns:
             A `WindowSet` object.
         """
+
     def explode(self):
         """Explodes an edge and returns all instances it had been updated as seperate edges"""
-    def explode_layers(self): ...
+
+    def explode_layers(self):
+        ...
+
     def has_layer(self, name):
-        """Check if Edges has the layer `"name"`"""
+        """ Check if Edges has the layer `"name"`"""
+
     def history(self):
         """
         Returns all timestamps of edges, when an edge is added or change to an edge is made.
@@ -681,6 +775,7 @@ class Edges:
            A list of lists unix timestamps.
 
         """
+
     def history_date_time(self):
         """
         Returns all timestamps of edges, when an edge is added or change to an edge is made.
@@ -689,15 +784,20 @@ class Edges:
            A list of lists of timestamps.
 
         """
+
     @property
     def id(self):
         """Returns all ids of the edges."""
+
     def is_deleted(self):
         """Check if the edges are deleted"""
+
     def is_self_loop(self):
         """Check if the edges are on the same node"""
+
     def is_valid(self):
         """Check if the edges are valid (i.e. not deleted)"""
+
     @property
     def latest_date_time(self):
         """
@@ -706,6 +806,7 @@ class Edges:
         Returns:
           Latest date time of the edges.
         """
+
     @property
     def latest_time(self):
         """
@@ -714,6 +815,7 @@ class Edges:
         Returns:
          Latest time of the edges.
         """
+
     def layer(self, name):
         """
          Return a view of Edges containing the layer `"name"`
@@ -722,6 +824,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     @property
     def layer_name(self):
         """
@@ -730,6 +833,7 @@ class Edges:
         Returns:
          The name of the layer
         """
+
     @property
     def layer_names(self):
         """
@@ -738,6 +842,7 @@ class Edges:
         Returns:
           A list of layer names
         """
+
     def layers(self, names):
         """
          Return a view of Edges containing all layers `names`
@@ -749,12 +854,15 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     @property
     def nbr(self):
         """Returns the node at the other end of the edge (same as `dst()` for out-edges and `src()` for in-edges)"""
+
     @property
     def properties(self):
         """Returns all properties of the edges"""
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -768,6 +876,7 @@ class Edges:
         Returns:
             A `WindowSet` object.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -777,6 +886,7 @@ class Edges:
         Returns:
              A Edges object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -787,6 +897,7 @@ class Edges:
         Returns:
              A Edges object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -794,9 +905,11 @@ class Edges:
         Arguments:
 
         """
+
     @property
     def src(self):
         """Returns the source node of the edge."""
+
     @property
     def start(self):
         """
@@ -805,6 +918,7 @@ class Edges:
         Returns:
             The earliest time that this Edges is valid or None if the Edges is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -813,6 +927,7 @@ class Edges:
         Returns:
              The earliest datetime that this Edges is valid or None if the Edges is valid for all times.
         """
+
     @property
     def time(self):
         """
@@ -821,9 +936,8 @@ class Edges:
         Returns:
           Time of edge
         """
-    def to_df(
-        self, include_property_history=True, convert_datetime=False, explode=False
-    ):
+
+    def to_df(self, include_property_history=True, convert_datetime=False, explode=False):
         """
         Converts the graph's edges into a Pandas DataFrame.
 
@@ -842,6 +956,7 @@ class Edges:
         Returns:
             If successful, this PyObject will be a Pandas DataFrame.
         """
+
     def valid_layers(self, names):
         """
          Return a view of Edges containing all layers `names`
@@ -853,6 +968,7 @@ class Edges:
         Returns:
              Edges: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Edges including all events between `start` (inclusive) and `end` (exclusive)
@@ -864,15 +980,17 @@ class Edges:
         Returns:
         r    A Edges object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Edges"""
+        """ Get the window size (difference between start and end) for this Edges"""
 
 class Graph:
     """A temporal graph."""
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def add_constant_properties(self, properties):
         """
         Adds static properties to the graph.
@@ -883,6 +1001,7 @@ class Graph:
         Returns:
            None
         """
+
     def add_edge(self, timestamp, src, dst, properties=None, layer=None):
         """
         Adds a new edge with the given source and destination nodes and properties to the graph.
@@ -897,6 +1016,7 @@ class Graph:
         Returns:
           None
         """
+
     def add_node(self, timestamp, id, properties=None, node_type=None):
         """
         Adds a new node with the given id and properties to the graph.
@@ -909,6 +1029,7 @@ class Graph:
         Returns:
           None
         """
+
     def add_property(self, timestamp, properties):
         """
         Adds properties to the graph.
@@ -920,6 +1041,7 @@ class Graph:
         Returns:
            None
         """
+
     def after(self, start):
         """
          Create a view of the GraphView including all events after `start` (exclusive).
@@ -930,6 +1052,7 @@ class Graph:
         Returns:
              A GraphView object.
         """
+
     def at(self, time):
         """
          Create a view of the GraphView including all events at `time`.
@@ -940,6 +1063,7 @@ class Graph:
         Returns:
              A GraphView object.
         """
+
     def before(self, end):
         """
          Create a view of the GraphView including all events before `end` (exclusive).
@@ -950,6 +1074,7 @@ class Graph:
         Returns:
              A GraphView object.
         """
+
     def cache(self, path):
         """
          Write Graph to cache file and initialise the cache.
@@ -960,6 +1085,7 @@ class Graph:
         Arguments:
             path (str): The path to the cache file
         """
+
     def count_edges(self):
         """
         Number of edges in the graph
@@ -967,6 +1093,7 @@ class Graph:
         Returns:
            the number of edges in the graph
         """
+
     def count_nodes(self):
         """
         Number of nodes in the graph
@@ -974,6 +1101,7 @@ class Graph:
         Returns:
           the number of nodes in the graph
         """
+
     def count_temporal_edges(self):
         """
         Number of edges in the graph
@@ -981,12 +1109,14 @@ class Graph:
         Returns:
            the number of temporal edges in the graph
         """
+
     def default_layer(self):
         """
          Return a view of GraphView containing only the default edge layer
         Returns:
              GraphView: The layered view
         """
+
     @staticmethod
     def deserialise(bytes):
         """
@@ -998,6 +1128,7 @@ class Graph:
         Returns:
            Graph
         """
+
     @property
     def earliest_date_time(self):
         """
@@ -1006,6 +1137,7 @@ class Graph:
         Returns:
             the datetime of the earliest activity in the graph
         """
+
     @property
     def earliest_time(self):
         """
@@ -1014,6 +1146,7 @@ class Graph:
         Returns:
             the timestamp of the earliest activity in the graph
         """
+
     def edge(self, src, dst):
         """
         Gets the edge with the specified source and destination nodes
@@ -1025,6 +1158,7 @@ class Graph:
         Returns:
             the edge with the specified source and destination nodes, or None if the edge does not exist
         """
+
     @property
     def edges(self):
         """
@@ -1033,6 +1167,7 @@ class Graph:
         Returns:
          the edges in the graph
         """
+
     @property
     def end(self):
         """
@@ -1041,6 +1176,7 @@ class Graph:
         Returns:
            The latest time that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -1049,6 +1185,7 @@ class Graph:
         Returns:
              The latest datetime that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of GraphView containing all layers except the excluded `name`
@@ -1060,6 +1197,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of GraphView containing all layers except the excluded `names`
@@ -1071,6 +1209,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_nodes(self, nodes):
         """
         Returns a subgraph given a set of nodes that are excluded from the subgraph
@@ -1081,6 +1220,7 @@ class Graph:
         Returns:
            GraphView - Returns the subgraph
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of GraphView containing all layers except the excluded `name`
@@ -1090,6 +1230,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of GraphView containing all layers except the excluded `names`
@@ -1099,6 +1240,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -1111,6 +1253,7 @@ class Graph:
         Returns:
             A `WindowSet` object.
         """
+
     def find_edges(self, properties_dict):
         """
         Get the edges that match the properties name and value
@@ -1119,6 +1262,7 @@ class Graph:
         Returns:
            the edges that match the properties name and value
         """
+
     def find_nodes(self, properties_dict):
         """
         Get the nodes that match the properties name and value
@@ -1127,6 +1271,7 @@ class Graph:
         Returns:
            the nodes that match the properties name and value
         """
+
     def get_all_node_types(self):
         """
         Returns all the node types in the graph.
@@ -1134,6 +1279,7 @@ class Graph:
         Returns:
         A list of node types
         """
+
     def has_edge(self, src, dst):
         """
         Returns true if the graph contains the specified edge
@@ -1145,8 +1291,10 @@ class Graph:
         Returns:
          true if the graph contains the specified edge, false otherwise
         """
+
     def has_layer(self, name):
-        """Check if GraphView has the layer `"name"`"""
+        """ Check if GraphView has the layer `"name"`"""
+
     def has_node(self, id):
         """
         Returns true if the graph contains the specified node
@@ -1157,6 +1305,7 @@ class Graph:
         Returns:
           true if the graph contains the specified node, false otherwise
         """
+
     def import_edge(self, edge, force=False):
         """
         Import a single edge into the graph.
@@ -1172,6 +1321,7 @@ class Graph:
         Returns:
             Result<EdgeView<Graph, Graph>, GraphError> - A Result object which is Ok if the edge was successfully imported, and Err otherwise.
         """
+
     def import_edges(self, edges, force=False):
         """
         Import multiple edges into the graph.
@@ -1185,6 +1335,7 @@ class Graph:
             force (boolean) - An optional boolean flag indicating whether to force the import of the edges.
 
         """
+
     def import_node(self, node, force=False):
         """
         Import a single node into the graph.
@@ -1199,6 +1350,7 @@ class Graph:
         Returns:
             Result<NodeView<Graph, Graph>, GraphError> - A Result object which is Ok if the node was successfully imported, and Err otherwise.
         """
+
     def import_nodes(self, nodes, force=False):
         """
         Import multiple nodes into the graph.
@@ -1212,6 +1364,7 @@ class Graph:
             force (boolean) - An optional boolean flag indicating whether to force the import of the nodes.
 
         """
+
     def index(self):
         """
         Indexes all node and edge properties.
@@ -1221,6 +1374,7 @@ class Graph:
         Returns:
            GraphIndex - Returns a GraphIndex
         """
+
     def largest_connected_component(self):
         """
         Gives the large connected component of a graph.
@@ -1232,6 +1386,7 @@ class Graph:
         A raphtory graph, which essentially is a sub-graph of the graph `g`
 
         """
+
     @property
     def latest_date_time(self):
         """
@@ -1240,6 +1395,7 @@ class Graph:
         Returns:
             the datetime of the latest activity in the graph
         """
+
     @property
     def latest_time(self):
         """
@@ -1248,6 +1404,7 @@ class Graph:
         Returns:
             the timestamp of the latest activity in the graph
         """
+
     def layer(self, name):
         """
          Return a view of GraphView containing the layer `"name"`
@@ -1256,6 +1413,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     def layers(self, names):
         """
          Return a view of GraphView containing all layers `names`
@@ -1267,6 +1425,7 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
+
     @staticmethod
     def load_cached(path):
         """
@@ -1281,16 +1440,8 @@ class Graph:
         Returns:
            Graph
         """
-    def load_edge_props_from_pandas(
-        self,
-        df,
-        src,
-        dst,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edge_props_from_pandas(self, df, src, dst, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edge properties from a Pandas DataFrame.
 
@@ -1309,16 +1460,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edge_props_from_parquet(
-        self,
-        parquet_path,
-        src,
-        dst,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edge_props_from_parquet(self, parquet_path, src, dst, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edge properties from parquet file
 
@@ -1337,18 +1480,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edges_from_pandas(
-        self,
-        df,
-        time,
-        src,
-        dst,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edges_from_pandas(self, df, time, src, dst, properties=None, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edges from a Pandas DataFrame into the graph.
 
@@ -1368,18 +1501,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edges_from_parquet(
-        self,
-        parquet_path,
-        time,
-        src,
-        dst,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edges_from_parquet(self, parquet_path, time, src, dst, properties=None, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edges from a Parquet file into the graph.
 
@@ -1399,6 +1522,7 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
+
     @staticmethod
     def load_from_file(path):
         """
@@ -1410,15 +1534,8 @@ class Graph:
         Returns:
            Graph
         """
-    def load_node_props_from_pandas(
-        self,
-        df,
-        id,
-        node_type=None,
-        node_type_col=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_node_props_from_pandas(self, df, id, node_type=None, node_type_col=None, constant_properties=None, shared_constant_properties=None):
         """
         Load node properties from a Pandas DataFrame.
 
@@ -1436,15 +1553,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_node_props_from_parquet(
-        self,
-        parquet_path,
-        id,
-        node_type=None,
-        node_type_col=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_node_props_from_parquet(self, parquet_path, id, node_type=None, node_type_col=None, constant_properties=None, shared_constant_properties=None):
         """
         Load node properties from a parquet file.
 
@@ -1462,17 +1572,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_nodes_from_pandas(
-        self,
-        df,
-        time,
-        id,
-        node_type=None,
-        node_type_col=None,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_nodes_from_pandas(self, df, time, id, node_type=None, node_type_col=None, properties=None, constant_properties=None, shared_constant_properties=None):
         """
         Load nodes from a Pandas DataFrame into the graph.
 
@@ -1491,17 +1592,8 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_nodes_from_parquet(
-        self,
-        parquet_path,
-        time,
-        id,
-        node_type=None,
-        node_type_col=None,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_nodes_from_parquet(self, parquet_path, time, id, node_type=None, node_type_col=None, properties=None, constant_properties=None, shared_constant_properties=None):
         """
         Load nodes from a Parquet file into the graph.
 
@@ -1520,6 +1612,7 @@ class Graph:
         Raises:
             GraphError: If the operation fails.
         """
+
     def materialize(self):
         """
         Returns a 'materialized' clone of the graph view - i.e. a new graph with a copy of the data seen within the view instead of just a mask over the original graph
@@ -1527,6 +1620,7 @@ class Graph:
         Returns:
            GraphView - Returns a graph clone
         """
+
     def node(self, id):
         """
         Gets the node with the specified id
@@ -1537,6 +1631,7 @@ class Graph:
         Returns:
           the node with the specified id, or None if the node does not exist
         """
+
     @property
     def nodes(self):
         """
@@ -1545,10 +1640,13 @@ class Graph:
         Returns:
          the nodes in the graph
         """
+
     def persist_as_disk_graph(self, graph_dir):
         """save graph in disk_graph format and memory map the result"""
+
     def persistent_graph(self):
         """Get persistent graph"""
+
     @property
     def properties(self):
         """
@@ -1558,6 +1656,7 @@ class Graph:
         Returns:
            HashMap<String, Prop> - Properties paired with their names
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -1571,6 +1670,7 @@ class Graph:
         Returns:
             A `WindowSet` object.
         """
+
     def save_to_file(self, path):
         """
          Saves the Graph to the given path.
@@ -1578,6 +1678,7 @@ class Graph:
         Arguments:
          path (str): The path to the file.
         """
+
     def serialise(self):
         """
          Serialise Graph to bytes.
@@ -1585,6 +1686,7 @@ class Graph:
         Returns:
           Bytes
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -1594,6 +1696,7 @@ class Graph:
         Returns:
              A GraphView object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -1604,6 +1707,7 @@ class Graph:
         Returns:
              A GraphView object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -1611,6 +1715,7 @@ class Graph:
         Arguments:
 
         """
+
     @property
     def start(self):
         """
@@ -1619,6 +1724,7 @@ class Graph:
         Returns:
             The earliest time that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -1627,6 +1733,7 @@ class Graph:
         Returns:
              The earliest datetime that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     def subgraph(self, nodes):
         """
         Returns a subgraph given a set of nodes
@@ -1637,6 +1744,7 @@ class Graph:
         Returns:
            GraphView - Returns the subgraph
         """
+
     def subgraph_node_types(self, node_types):
         """
         Returns a subgraph filtered by node types given a set of node types
@@ -1647,15 +1755,11 @@ class Graph:
         Returns:
            GraphView - Returns the subgraph
         """
-    def to_disk_graph(self, graph_dir): ...
-    def to_networkx(
-        self,
-        explode_edges=False,
-        include_node_properties=True,
-        include_edge_properties=True,
-        include_update_history=True,
-        include_property_history=True,
-    ):
+
+    def to_disk_graph(self, graph_dir):
+        ...
+
+    def to_networkx(self, explode_edges=False, include_node_properties=True, include_edge_properties=True, include_update_history=True, include_property_history=True):
         """
         Returns a graph with NetworkX.
 
@@ -1673,18 +1777,8 @@ class Graph:
             Returns:
                 A Networkx MultiDiGraph.
         """
-    def to_pyvis(
-        self,
-        explode_edges=False,
-        edge_color="#000000",
-        shape=None,
-        node_image=None,
-        edge_weight=None,
-        edge_label=None,
-        colour_nodes_by_type=False,
-        notebook=False,
-        **kwargs,
-    ):
+
+    def to_pyvis(self, explode_edges=False, edge_color="#000000", shape=None, node_image=None, edge_weight=None, edge_label=None, colour_nodes_by_type=False, notebook=False, **kwargs):
         """
         Draw a graph with PyVis.
         Pyvis is a required dependency. If you intend to use this function make sure that you install Pyvis
@@ -1708,9 +1802,11 @@ class Graph:
             Returns:
                 A pyvis network
         """
+
     @property
     def unique_layers(self):
         """Return all the layer ids in the graph"""
+
     def update_constant_properties(self, properties):
         """
         Updates static properties to the graph.
@@ -1721,6 +1817,7 @@ class Graph:
         Returns:
            None
         """
+
     def valid_layers(self, names):
         """
          Return a view of GraphView containing all layers `names`
@@ -1732,16 +1829,8 @@ class Graph:
         Returns:
              GraphView: The layered view
         """
-    def vectorise(
-        self,
-        embedding,
-        cache=None,
-        overwrite_cache=False,
-        graph_document=None,
-        node_document=None,
-        edge_document=None,
-        verbose=False,
-    ):
+
+    def vectorise(self, embedding, cache=None, overwrite_cache=False, graph_document=None, node_document=None, edge_document=None, verbose=False):
         """
         Create a VectorisedGraph from the current graph
 
@@ -1756,6 +1845,7 @@ class Graph:
         Returns:
           A VectorisedGraph with all the documents/embeddings computed and with an initial empty selection
         """
+
     def window(self, start, end):
         """
          Create a view of the GraphView including all events between `start` (inclusive) and `end` (exclusive)
@@ -1767,9 +1857,11 @@ class Graph:
         Returns:
         r    A GraphView object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this GraphView"""
+        """ Get the window size (difference between start and end) for this GraphView"""
+
     def write_updates(self):
         """Persist the new updates by appending them to the cache file."""
 
@@ -1782,9 +1874,8 @@ class GraphIndex:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
-    def fuzzy_search_edges(
-        self, query, limit=25, offset=0, prefix=False, levenshtein_distance=0
-    ):
+
+    def fuzzy_search_edges(self, query, limit=25, offset=0, prefix=False, levenshtein_distance=0):
         """
         Searches for edges which match the given query. This uses Tantivy's fuzzy search.
 
@@ -1798,9 +1889,8 @@ class GraphIndex:
         Returns:
            A list of edges which match the query. The list will be empty if no edges match the query.
         """
-    def fuzzy_search_nodes(
-        self, query, limit=25, offset=0, prefix=False, levenshtein_distance=0
-    ):
+
+    def fuzzy_search_nodes(self, query, limit=25, offset=0, prefix=False, levenshtein_distance=0):
         """
         Searches for nodes which match the given query. This uses Tantivy's fuzzy search.
         If you would like to better understand the query syntax, please visit our documentation at https://docs.raphtory.com
@@ -1815,6 +1905,7 @@ class GraphIndex:
         Returns:
            A list of nodes which match the query. The list will be empty if no nodes match.
         """
+
     def search_edges(self, query, limit=25, offset=0):
         """
         Searches for edges which match the given query. This uses Tantivy's exact search.
@@ -1827,6 +1918,7 @@ class GraphIndex:
         Returns:
            A list of edges which match the query. The list will be empty if no edges match the query.
         """
+
     def search_nodes(self, query, limit=25, offset=0):
         """
         Searches for nodes which match the given query. This uses Tantivy's exact search.
@@ -1841,8 +1933,10 @@ class GraphIndex:
         """
 
 class MutableEdge:
+
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def add_constant_properties(self, properties, layer=None):
         """
         Add constant properties to an edge in the graph.
@@ -1858,6 +1952,7 @@ class MutableEdge:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def add_updates(self, t, properties=None, layer=None):
         """
         Add updates to an edge in the graph at a specified time.
@@ -1872,6 +1967,7 @@ class MutableEdge:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def after(self, start):
         """
          Create a view of the Edge including all events after `start` (exclusive).
@@ -1882,6 +1978,7 @@ class MutableEdge:
         Returns:
              A Edge object.
         """
+
     def at(self, time):
         """
          Create a view of the Edge including all events at `time`.
@@ -1892,6 +1989,7 @@ class MutableEdge:
         Returns:
              A Edge object.
         """
+
     def before(self, end):
         """
          Create a view of the Edge including all events before `end` (exclusive).
@@ -1902,6 +2000,7 @@ class MutableEdge:
         Returns:
              A Edge object.
         """
+
     @property
     def date_time(self):
         """
@@ -1910,12 +2009,14 @@ class MutableEdge:
         Returns:
             (datetime) the datetime of an exploded edge
         """
+
     def default_layer(self):
         """
          Return a view of Edge containing only the default edge layer
         Returns:
              Edge: The layered view
         """
+
     def deletions(self):
         """
         Returns a list of timestamps of when an edge is deleted
@@ -1923,6 +2024,7 @@ class MutableEdge:
         Returns:
             A list of unix timestamps
         """
+
     def deletions_data_time(self):
         """
         Returns a list of timestamps of when an edge is deleted
@@ -1930,9 +2032,11 @@ class MutableEdge:
         Returns:
             A list of DateTime objects
         """
+
     @property
     def dst(self):
         """Returns the destination node of the edge."""
+
     @property
     def earliest_date_time(self):
         """
@@ -1941,6 +2045,7 @@ class MutableEdge:
         Returns:
             the earliest datetime of an edge
         """
+
     @property
     def earliest_time(self):
         """
@@ -1949,6 +2054,7 @@ class MutableEdge:
         Returns:
             (int) The earliest time of an edge
         """
+
     @property
     def end(self):
         """
@@ -1957,6 +2063,7 @@ class MutableEdge:
         Returns:
            The latest time that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -1965,6 +2072,7 @@ class MutableEdge:
         Returns:
              The latest datetime that this Edge is valid or None if the Edge is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Edge containing all layers except the excluded `name`
@@ -1976,6 +2084,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Edge containing all layers except the excluded `names`
@@ -1987,6 +2096,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Edge containing all layers except the excluded `name`
@@ -1996,6 +2106,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Edge containing all layers except the excluded `names`
@@ -2005,6 +2116,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -2017,11 +2129,16 @@ class MutableEdge:
         Returns:
             A `WindowSet` object.
         """
+
     def explode(self):
         """Explodes an edge and returns all instances it had been updated as seperate edges"""
-    def explode_layers(self): ...
+
+    def explode_layers(self):
+        ...
+
     def has_layer(self, name):
-        """Check if Edge has the layer `"name"`"""
+        """ Check if Edge has the layer `"name"`"""
+
     def history(self):
         """
         Returns a list of timestamps of when an edge is added or change to an edge is made.
@@ -2030,6 +2147,7 @@ class MutableEdge:
             A list of unix timestamps.
 
         """
+
     def history_date_time(self):
         """
         Returns a list of timestamps of when an edge is added or change to an edge is made.
@@ -2038,15 +2156,20 @@ class MutableEdge:
             A list of timestamps.
 
         """
+
     @property
     def id(self):
         """The id of the edge."""
+
     def is_deleted(self):
         """Check if the edge is currently deleted"""
+
     def is_self_loop(self):
         """Check if the edge is on the same node"""
+
     def is_valid(self):
         """Check if the edge is currently valid (i.e., not deleted)"""
+
     @property
     def latest_date_time(self):
         """
@@ -2055,6 +2178,7 @@ class MutableEdge:
         Returns:
             (datetime) the latest datetime of an edge
         """
+
     @property
     def latest_time(self):
         """
@@ -2063,6 +2187,7 @@ class MutableEdge:
         Returns:
             (int) The latest time of an edge
         """
+
     def layer(self, name):
         """
          Return a view of Edge containing the layer `"name"`
@@ -2071,6 +2196,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     @property
     def layer_name(self):
         """
@@ -2079,6 +2205,7 @@ class MutableEdge:
         Returns:
             (List<str>) The name of the layer
         """
+
     @property
     def layer_names(self):
         """
@@ -2087,6 +2214,7 @@ class MutableEdge:
         Returns:
             (List<str>) The name of the layer
         """
+
     def layers(self, names):
         """
          Return a view of Edge containing all layers `names`
@@ -2098,9 +2226,11 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     @property
     def nbr(self):
         """Returns the node at the other end of the edge (same as `dst()` for out-edges and `src()` for in-edges)"""
+
     @property
     def properties(self):
         """
@@ -2109,6 +2239,7 @@ class MutableEdge:
         Returns:
           Properties on the Edge.
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -2122,6 +2253,7 @@ class MutableEdge:
         Returns:
             A `WindowSet` object.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -2131,6 +2263,7 @@ class MutableEdge:
         Returns:
              A Edge object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -2141,6 +2274,7 @@ class MutableEdge:
         Returns:
              A Edge object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -2148,9 +2282,11 @@ class MutableEdge:
         Arguments:
 
         """
+
     @property
     def src(self):
         """Returns the source node of the edge."""
+
     @property
     def start(self):
         """
@@ -2159,6 +2295,7 @@ class MutableEdge:
         Returns:
             The earliest time that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -2167,6 +2304,7 @@ class MutableEdge:
         Returns:
              The earliest datetime that this Edge is valid or None if the Edge is valid for all times.
         """
+
     @property
     def time(self):
         """
@@ -2175,6 +2313,7 @@ class MutableEdge:
         Returns:
             (int) The time of an exploded edge
         """
+
     def update_constant_properties(self, properties, layer=None):
         """
         Update constant properties of an edge in the graph overwriting existing values.
@@ -2190,6 +2329,7 @@ class MutableEdge:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def valid_layers(self, names):
         """
          Return a view of Edge containing all layers `names`
@@ -2201,6 +2341,7 @@ class MutableEdge:
         Returns:
              Edge: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Edge including all events between `start` (inclusive) and `end` (exclusive)
@@ -2212,13 +2353,16 @@ class MutableEdge:
         Returns:
         r    A Edge object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Edge"""
+        """ Get the window size (difference between start and end) for this Edge"""
 
 class MutableNode:
+
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def add_constant_properties(self, properties):
         """
         Add constant properties to a node in the graph.
@@ -2233,6 +2377,7 @@ class MutableNode:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def add_updates(self, t, properties=None):
         """
         Add updates to a node in the graph at a specified time.
@@ -2247,6 +2392,7 @@ class MutableNode:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def after(self, start):
         """
          Create a view of the Node including all events after `start` (exclusive).
@@ -2257,6 +2403,7 @@ class MutableNode:
         Returns:
              A Node object.
         """
+
     def at(self, time):
         """
          Create a view of the Node including all events at `time`.
@@ -2267,6 +2414,7 @@ class MutableNode:
         Returns:
              A Node object.
         """
+
     def before(self, end):
         """
          Create a view of the Node including all events before `end` (exclusive).
@@ -2277,12 +2425,14 @@ class MutableNode:
         Returns:
              A Node object.
         """
+
     def default_layer(self):
         """
          Return a view of Node containing only the default edge layer
         Returns:
              Node: The layered view
         """
+
     def degree(self):
         """
         Get the degree of this node (i.e., the number of edges that are incident to it).
@@ -2290,6 +2440,7 @@ class MutableNode:
         Returns
             The degree of this node.
         """
+
     @property
     def earliest_date_time(self):
         """
@@ -2298,6 +2449,7 @@ class MutableNode:
         Returns:
             The earliest datetime that the node exists as an integer.
         """
+
     @property
     def earliest_time(self):
         """
@@ -2306,6 +2458,7 @@ class MutableNode:
         Returns:
             The earliest time that the node exists as an integer.
         """
+
     @property
     def edges(self):
         """
@@ -2315,6 +2468,7 @@ class MutableNode:
 
         An iterator over the edges that are incident to this node.
         """
+
     @property
     def end(self):
         """
@@ -2323,6 +2477,7 @@ class MutableNode:
         Returns:
            The latest time that this Node is valid or None if the Node is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -2331,6 +2486,7 @@ class MutableNode:
         Returns:
              The latest datetime that this Node is valid or None if the Node is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Node containing all layers except the excluded `name`
@@ -2342,6 +2498,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Node containing all layers except the excluded `names`
@@ -2353,6 +2510,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Node containing all layers except the excluded `name`
@@ -2362,6 +2520,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Node containing all layers except the excluded `names`
@@ -2371,6 +2530,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -2383,8 +2543,10 @@ class MutableNode:
         Returns:
             A `WindowSet` object.
         """
+
     def has_layer(self, name):
-        """Check if Node has the layer `"name"`"""
+        """ Check if Node has the layer `"name"`"""
+
     def history(self):
         """
         Returns the history of a node, including node additions and changes made to node.
@@ -2392,6 +2554,7 @@ class MutableNode:
         Returns:
             A list of unix timestamps of the event history of node.
         """
+
     def history_date_time(self):
         """
         Returns the history of a node, including node additions and changes made to node.
@@ -2400,6 +2563,7 @@ class MutableNode:
             A list of timestamps of the event history of node.
 
         """
+
     @property
     def id(self):
         """
@@ -2409,6 +2573,7 @@ class MutableNode:
         Returns:
            The id of the node as an integer.
         """
+
     def in_degree(self):
         """
         Get the in-degree of this node (i.e., the number of edges that are incident to it from other nodes).
@@ -2416,6 +2581,7 @@ class MutableNode:
         Returns:
            The in-degree of this node.
         """
+
     @property
     def in_edges(self):
         """
@@ -2425,6 +2591,7 @@ class MutableNode:
 
         An iterator over the edges that point into this node.
         """
+
     @property
     def in_neighbours(self):
         """
@@ -2434,6 +2601,7 @@ class MutableNode:
 
         An iterator over the neighbours of this node that point into this node.
         """
+
     @property
     def latest_date_time(self):
         """
@@ -2445,6 +2613,7 @@ class MutableNode:
         Returns:
             The latest datetime that the node exists as an integer.
         """
+
     @property
     def latest_time(self):
         """
@@ -2453,6 +2622,7 @@ class MutableNode:
         Returns:
             The latest time that the node exists as an integer.
         """
+
     def layer(self, name):
         """
          Return a view of Node containing the layer `"name"`
@@ -2461,6 +2631,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def layers(self, names):
         """
          Return a view of Node containing all layers `names`
@@ -2472,6 +2643,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     @property
     def name(self):
         """
@@ -2480,6 +2652,7 @@ class MutableNode:
         Returns:
             The name of the node as a string.
         """
+
     @property
     def neighbours(self):
         """
@@ -2489,9 +2662,11 @@ class MutableNode:
 
         An iterator over the neighbours of this node.
         """
+
     @property
     def node_type(self):
         """Returns the type of node"""
+
     def out_degree(self):
         """
         Get the out-degree of this node (i.e., the number of edges that are incident to it from this node).
@@ -2499,6 +2674,7 @@ class MutableNode:
         Returns:
           The out-degree of this node.
         """
+
     @property
     def out_edges(self):
         """
@@ -2508,6 +2684,7 @@ class MutableNode:
 
         An iterator over the edges that point out of this node.
         """
+
     @property
     def out_neighbours(self):
         """
@@ -2517,6 +2694,7 @@ class MutableNode:
 
         An iterator over the neighbours of this node that point out of this node.
         """
+
     @property
     def properties(self):
         """
@@ -2525,6 +2703,7 @@ class MutableNode:
         Returns:
             A list of properties.
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -2538,6 +2717,7 @@ class MutableNode:
         Returns:
             A `WindowSet` object.
         """
+
     def set_node_type(self, new_type):
         """
         Set the type on the node. This only works if the type has not been previously set, otherwise will
@@ -2549,6 +2729,7 @@ class MutableNode:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -2558,6 +2739,7 @@ class MutableNode:
         Returns:
              A Node object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -2568,6 +2750,7 @@ class MutableNode:
         Returns:
              A Node object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -2575,6 +2758,7 @@ class MutableNode:
         Arguments:
 
         """
+
     @property
     def start(self):
         """
@@ -2583,6 +2767,7 @@ class MutableNode:
         Returns:
             The earliest time that this Node is valid or None if the Node is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -2591,6 +2776,7 @@ class MutableNode:
         Returns:
              The earliest datetime that this Node is valid or None if the Node is valid for all times.
         """
+
     def update_constant_properties(self, properties):
         """
         Update constant properties of a node in the graph overwriting existing values.
@@ -2605,6 +2791,7 @@ class MutableNode:
         Returns:
             Result: A result object indicating success or failure. On failure, it contains a GraphError.
         """
+
     def valid_layers(self, names):
         """
          Return a view of Node containing all layers `names`
@@ -2616,6 +2803,7 @@ class MutableNode:
         Returns:
              Node: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Node including all events between `start` (inclusive) and `end` (exclusive)
@@ -2627,15 +2815,17 @@ class MutableNode:
         Returns:
         r    A Node object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Node"""
+        """ Get the window size (difference between start and end) for this Node"""
 
 class Node:
     """A node (or node) in the graph."""
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def after(self, start):
         """
          Create a view of the Node including all events after `start` (exclusive).
@@ -2646,6 +2836,7 @@ class Node:
         Returns:
              A Node object.
         """
+
     def at(self, time):
         """
          Create a view of the Node including all events at `time`.
@@ -2656,6 +2847,7 @@ class Node:
         Returns:
              A Node object.
         """
+
     def before(self, end):
         """
          Create a view of the Node including all events before `end` (exclusive).
@@ -2666,12 +2858,14 @@ class Node:
         Returns:
              A Node object.
         """
+
     def default_layer(self):
         """
          Return a view of Node containing only the default edge layer
         Returns:
              Node: The layered view
         """
+
     def degree(self):
         """
         Get the degree of this node (i.e., the number of edges that are incident to it).
@@ -2679,6 +2873,7 @@ class Node:
         Returns
             The degree of this node.
         """
+
     @property
     def earliest_date_time(self):
         """
@@ -2687,6 +2882,7 @@ class Node:
         Returns:
             The earliest datetime that the node exists as an integer.
         """
+
     @property
     def earliest_time(self):
         """
@@ -2695,6 +2891,7 @@ class Node:
         Returns:
             The earliest time that the node exists as an integer.
         """
+
     @property
     def edges(self):
         """
@@ -2704,6 +2901,7 @@ class Node:
 
         An iterator over the edges that are incident to this node.
         """
+
     @property
     def end(self):
         """
@@ -2712,6 +2910,7 @@ class Node:
         Returns:
            The latest time that this Node is valid or None if the Node is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -2720,6 +2919,7 @@ class Node:
         Returns:
              The latest datetime that this Node is valid or None if the Node is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Node containing all layers except the excluded `name`
@@ -2731,6 +2931,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Node containing all layers except the excluded `names`
@@ -2742,6 +2943,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Node containing all layers except the excluded `name`
@@ -2751,6 +2953,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Node containing all layers except the excluded `names`
@@ -2760,6 +2963,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -2772,8 +2976,10 @@ class Node:
         Returns:
             A `WindowSet` object.
         """
+
     def has_layer(self, name):
-        """Check if Node has the layer `"name"`"""
+        """ Check if Node has the layer `"name"`"""
+
     def history(self):
         """
         Returns the history of a node, including node additions and changes made to node.
@@ -2781,6 +2987,7 @@ class Node:
         Returns:
             A list of unix timestamps of the event history of node.
         """
+
     def history_date_time(self):
         """
         Returns the history of a node, including node additions and changes made to node.
@@ -2789,6 +2996,7 @@ class Node:
             A list of timestamps of the event history of node.
 
         """
+
     @property
     def id(self):
         """
@@ -2798,6 +3006,7 @@ class Node:
         Returns:
            The id of the node as an integer.
         """
+
     def in_degree(self):
         """
         Get the in-degree of this node (i.e., the number of edges that are incident to it from other nodes).
@@ -2805,6 +3014,7 @@ class Node:
         Returns:
            The in-degree of this node.
         """
+
     @property
     def in_edges(self):
         """
@@ -2814,6 +3024,7 @@ class Node:
 
         An iterator over the edges that point into this node.
         """
+
     @property
     def in_neighbours(self):
         """
@@ -2823,6 +3034,7 @@ class Node:
 
         An iterator over the neighbours of this node that point into this node.
         """
+
     @property
     def latest_date_time(self):
         """
@@ -2834,6 +3046,7 @@ class Node:
         Returns:
             The latest datetime that the node exists as an integer.
         """
+
     @property
     def latest_time(self):
         """
@@ -2842,6 +3055,7 @@ class Node:
         Returns:
             The latest time that the node exists as an integer.
         """
+
     def layer(self, name):
         """
          Return a view of Node containing the layer `"name"`
@@ -2850,6 +3064,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def layers(self, names):
         """
          Return a view of Node containing all layers `names`
@@ -2861,6 +3076,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     @property
     def name(self):
         """
@@ -2869,6 +3085,7 @@ class Node:
         Returns:
             The name of the node as a string.
         """
+
     @property
     def neighbours(self):
         """
@@ -2878,9 +3095,11 @@ class Node:
 
         An iterator over the neighbours of this node.
         """
+
     @property
     def node_type(self):
         """Returns the type of node"""
+
     def out_degree(self):
         """
         Get the out-degree of this node (i.e., the number of edges that are incident to it from this node).
@@ -2888,6 +3107,7 @@ class Node:
         Returns:
           The out-degree of this node.
         """
+
     @property
     def out_edges(self):
         """
@@ -2897,6 +3117,7 @@ class Node:
 
         An iterator over the edges that point out of this node.
         """
+
     @property
     def out_neighbours(self):
         """
@@ -2906,6 +3127,7 @@ class Node:
 
         An iterator over the neighbours of this node that point out of this node.
         """
+
     @property
     def properties(self):
         """
@@ -2914,6 +3136,7 @@ class Node:
         Returns:
             A list of properties.
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -2927,6 +3150,7 @@ class Node:
         Returns:
             A `WindowSet` object.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -2936,6 +3160,7 @@ class Node:
         Returns:
              A Node object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -2946,6 +3171,7 @@ class Node:
         Returns:
              A Node object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -2953,6 +3179,7 @@ class Node:
         Arguments:
 
         """
+
     @property
     def start(self):
         """
@@ -2961,6 +3188,7 @@ class Node:
         Returns:
             The earliest time that this Node is valid or None if the Node is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -2969,6 +3197,7 @@ class Node:
         Returns:
              The earliest datetime that this Node is valid or None if the Node is valid for all times.
         """
+
     def valid_layers(self, names):
         """
          Return a view of Node containing all layers `names`
@@ -2980,6 +3209,7 @@ class Node:
         Returns:
              Node: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Node including all events between `start` (inclusive) and `end` (exclusive)
@@ -2991,15 +3221,17 @@ class Node:
         Returns:
         r    A Node object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Node"""
+        """ Get the window size (difference between start and end) for this Node"""
 
 class Nodes:
     """A list of nodes that can be iterated over."""
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def after(self, start):
         """
          Create a view of the Nodes including all events after `start` (exclusive).
@@ -3010,6 +3242,7 @@ class Nodes:
         Returns:
              A Nodes object.
         """
+
     def at(self, time):
         """
          Create a view of the Nodes including all events at `time`.
@@ -3020,6 +3253,7 @@ class Nodes:
         Returns:
              A Nodes object.
         """
+
     def before(self, end):
         """
          Create a view of the Nodes including all events before `end` (exclusive).
@@ -3030,6 +3264,7 @@ class Nodes:
         Returns:
              A Nodes object.
         """
+
     def collect(self):
         """
          Collect all nodes into a list
@@ -3037,12 +3272,14 @@ class Nodes:
         Returns:
              list[Node]: the list of nodes
         """
+
     def default_layer(self):
         """
          Return a view of Nodes containing only the default edge layer
         Returns:
              Nodes: The layered view
         """
+
     def degree(self):
         """
         Returns the number of edges of the nodes
@@ -3050,6 +3287,7 @@ class Nodes:
         Returns:
             An iterator of the number of edges of the nodes
         """
+
     @property
     def earliest_date_time(self):
         """
@@ -3058,9 +3296,11 @@ class Nodes:
         Returns:
         Earliest time of the nodes.
         """
+
     @property
     def earliest_time(self):
         """Returns an iterator over the nodes earliest time"""
+
     @property
     def edges(self):
         """
@@ -3070,6 +3310,7 @@ class Nodes:
 
         An iterator over the edges that are incident to this node.
         """
+
     @property
     def end(self):
         """
@@ -3078,6 +3319,7 @@ class Nodes:
         Returns:
            The latest time that this Nodes is valid or None if the Nodes is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -3086,6 +3328,7 @@ class Nodes:
         Returns:
              The latest datetime that this Nodes is valid or None if the Nodes is valid for all times.
         """
+
     def exclude_layer(self, name):
         """
          Return a view of Nodes containing all layers except the excluded `name`
@@ -3097,6 +3340,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of Nodes containing all layers except the excluded `names`
@@ -3108,6 +3352,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of Nodes containing all layers except the excluded `name`
@@ -3117,6 +3362,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of Nodes containing all layers except the excluded `names`
@@ -3126,6 +3372,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -3138,8 +3385,10 @@ class Nodes:
         Returns:
             A `WindowSet` object.
         """
+
     def has_layer(self, name):
-        """Check if Nodes has the layer `"name"`"""
+        """ Check if Nodes has the layer `"name"`"""
+
     def history(self):
         """
         Returns all timestamps of nodes, when an node is added or change to an node is made.
@@ -3148,6 +3397,7 @@ class Nodes:
            A list of unix timestamps.
 
         """
+
     def history_date_time(self):
         """
         Returns all timestamps of nodes, when an node is added or change to an node is made.
@@ -3156,9 +3406,11 @@ class Nodes:
            An  list of timestamps.
 
         """
+
     @property
     def id(self):
         """Returns an iterator over the nodes ids"""
+
     def in_degree(self):
         """
         Returns the number of in edges of the nodes
@@ -3166,6 +3418,7 @@ class Nodes:
         Returns:
             An iterator of the number of in edges of the nodes
         """
+
     @property
     def in_edges(self):
         """
@@ -3175,6 +3428,7 @@ class Nodes:
 
         An iterator over the edges that point into this node.
         """
+
     @property
     def in_neighbours(self):
         """
@@ -3184,6 +3438,7 @@ class Nodes:
 
         An iterator over the neighbours of this node that point into this node.
         """
+
     @property
     def latest_date_time(self):
         """
@@ -3192,9 +3447,11 @@ class Nodes:
         Returns:
           Latest date time of the nodes.
         """
+
     @property
     def latest_time(self):
         """Returns an iterator over the nodes latest time"""
+
     def layer(self, name):
         """
          Return a view of Nodes containing the layer `"name"`
@@ -3203,6 +3460,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def layers(self, names):
         """
          Return a view of Nodes containing all layers `names`
@@ -3214,9 +3472,11 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     @property
     def name(self):
         """Returns an iterator over the nodes name"""
+
     @property
     def neighbours(self):
         """
@@ -3226,9 +3486,11 @@ class Nodes:
 
         An iterator over the neighbours of this node.
         """
+
     @property
     def node_type(self):
         """Returns the type of node"""
+
     def out_degree(self):
         """
         Returns the number of out edges of the nodes
@@ -3236,6 +3498,7 @@ class Nodes:
         Returns:
             An iterator of the number of out edges of the nodes
         """
+
     @property
     def out_edges(self):
         """
@@ -3245,6 +3508,7 @@ class Nodes:
 
         An iterator over the edges that point out of this node.
         """
+
     @property
     def out_neighbours(self):
         """
@@ -3254,6 +3518,7 @@ class Nodes:
 
         An iterator over the neighbours of this node that point out of this node.
         """
+
     @property
     def properties(self):
         """
@@ -3262,6 +3527,7 @@ class Nodes:
         Returns:
             A List of properties
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -3275,6 +3541,7 @@ class Nodes:
         Returns:
             A `WindowSet` object.
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -3284,6 +3551,7 @@ class Nodes:
         Returns:
              A Nodes object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -3294,6 +3562,7 @@ class Nodes:
         Returns:
              A Nodes object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -3301,6 +3570,7 @@ class Nodes:
         Arguments:
 
         """
+
     @property
     def start(self):
         """
@@ -3309,6 +3579,7 @@ class Nodes:
         Returns:
             The earliest time that this Nodes is valid or None if the Nodes is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -3317,6 +3588,7 @@ class Nodes:
         Returns:
              The earliest datetime that this Nodes is valid or None if the Nodes is valid for all times.
         """
+
     def to_df(self, include_property_history=False, convert_datetime=False):
         """
         Converts the graph's nodes into a Pandas DataFrame.
@@ -3333,7 +3605,10 @@ class Nodes:
         Returns:
             If successful, this PyObject will be a Pandas DataFrame.
         """
-    def type_filter(self, node_types): ...
+
+    def type_filter(self, node_types):
+        ...
+
     def valid_layers(self, names):
         """
          Return a view of Nodes containing all layers `names`
@@ -3345,6 +3620,7 @@ class Nodes:
         Returns:
              Nodes: The layered view
         """
+
     def window(self, start, end):
         """
          Create a view of the Nodes including all events between `start` (inclusive) and `end` (exclusive)
@@ -3356,15 +3632,17 @@ class Nodes:
         Returns:
         r    A Nodes object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this Nodes"""
+        """ Get the window size (difference between start and end) for this Nodes"""
 
 class PersistentGraph:
     """A temporal graph that allows edges and nodes to be deleted."""
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def add_constant_properties(self, properties):
         """
         Adds static properties to the graph.
@@ -3375,6 +3653,7 @@ class PersistentGraph:
         Returns:
            None
         """
+
     def add_edge(self, timestamp, src, dst, properties=None, layer=None):
         """
         Adds a new edge with the given source and destination nodes and properties to the graph.
@@ -3389,6 +3668,7 @@ class PersistentGraph:
         Returns:
           None
         """
+
     def add_node(self, timestamp, id, properties=None, node_type=None):
         """
         Adds a new node with the given id and properties to the graph.
@@ -3402,6 +3682,7 @@ class PersistentGraph:
         Returns:
           None
         """
+
     def add_property(self, timestamp, properties):
         """
         Adds properties to the graph.
@@ -3413,6 +3694,7 @@ class PersistentGraph:
         Returns:
            None
         """
+
     def after(self, start):
         """
          Create a view of the GraphView including all events after `start` (exclusive).
@@ -3423,6 +3705,7 @@ class PersistentGraph:
         Returns:
              A GraphView object.
         """
+
     def at(self, time):
         """
          Create a view of the GraphView including all events at `time`.
@@ -3433,6 +3716,7 @@ class PersistentGraph:
         Returns:
              A GraphView object.
         """
+
     def before(self, end):
         """
          Create a view of the GraphView including all events before `end` (exclusive).
@@ -3443,6 +3727,7 @@ class PersistentGraph:
         Returns:
              A GraphView object.
         """
+
     def cache(self, path):
         """
          Write PersistentGraph to cache file and initialise the cache.
@@ -3453,6 +3738,7 @@ class PersistentGraph:
         Arguments:
             path (str): The path to the cache file
         """
+
     def count_edges(self):
         """
         Number of edges in the graph
@@ -3460,6 +3746,7 @@ class PersistentGraph:
         Returns:
            the number of edges in the graph
         """
+
     def count_nodes(self):
         """
         Number of nodes in the graph
@@ -3467,6 +3754,7 @@ class PersistentGraph:
         Returns:
           the number of nodes in the graph
         """
+
     def count_temporal_edges(self):
         """
         Number of edges in the graph
@@ -3474,12 +3762,14 @@ class PersistentGraph:
         Returns:
            the number of temporal edges in the graph
         """
+
     def default_layer(self):
         """
          Return a view of GraphView containing only the default edge layer
         Returns:
              GraphView: The layered view
         """
+
     def delete_edge(self, timestamp, src, dst, layer=None):
         """
         Deletes an edge given the timestamp, src and dst nodes and layer (optional)
@@ -3493,6 +3783,7 @@ class PersistentGraph:
         Returns:
          The deleted edge
         """
+
     @staticmethod
     def deserialise(bytes):
         """
@@ -3504,6 +3795,7 @@ class PersistentGraph:
         Returns:
            PersistentGraph
         """
+
     @property
     def earliest_date_time(self):
         """
@@ -3512,6 +3804,7 @@ class PersistentGraph:
         Returns:
             the datetime of the earliest activity in the graph
         """
+
     @property
     def earliest_time(self):
         """
@@ -3520,6 +3813,7 @@ class PersistentGraph:
         Returns:
             the timestamp of the earliest activity in the graph
         """
+
     def edge(self, src, dst):
         """
         Gets the edge with the specified source and destination nodes
@@ -3531,6 +3825,7 @@ class PersistentGraph:
         Returns:
             the edge with the specified source and destination nodes, or None if the edge does not exist
         """
+
     @property
     def edges(self):
         """
@@ -3539,6 +3834,7 @@ class PersistentGraph:
         Returns:
          the edges in the graph
         """
+
     @property
     def end(self):
         """
@@ -3547,6 +3843,7 @@ class PersistentGraph:
         Returns:
            The latest time that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     @property
     def end_date_time(self):
         """
@@ -3555,8 +3852,10 @@ class PersistentGraph:
         Returns:
              The latest datetime that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     def event_graph(self):
         """Get event graph"""
+
     def exclude_layer(self, name):
         """
          Return a view of GraphView containing all layers except the excluded `name`
@@ -3568,6 +3867,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_layers(self, names):
         """
          Return a view of GraphView containing all layers except the excluded `names`
@@ -3579,6 +3879,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_nodes(self, nodes):
         """
         Returns a subgraph given a set of nodes that are excluded from the subgraph
@@ -3589,6 +3890,7 @@ class PersistentGraph:
         Returns:
            GraphView - Returns the subgraph
         """
+
     def exclude_valid_layer(self, name):
         """
          Return a view of GraphView containing all layers except the excluded `name`
@@ -3598,6 +3900,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     def exclude_valid_layers(self, names):
         """
          Return a view of GraphView containing all layers except the excluded `names`
@@ -3607,6 +3910,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     def expanding(self, step):
         """
         Creates a `WindowSet` with the given `step` size using an expanding window.
@@ -3619,6 +3923,7 @@ class PersistentGraph:
         Returns:
             A `WindowSet` object.
         """
+
     def find_edges(self, properties_dict):
         """
         Get the edges that match the properties name and value
@@ -3627,6 +3932,7 @@ class PersistentGraph:
         Returns:
            the edges that match the properties name and value
         """
+
     def find_nodes(self, properties_dict):
         """
         Get the nodes that match the properties name and value
@@ -3635,6 +3941,7 @@ class PersistentGraph:
         Returns:
            the nodes that match the properties name and value
         """
+
     def get_all_node_types(self):
         """
         Returns all the node types in the graph.
@@ -3642,6 +3949,7 @@ class PersistentGraph:
         Returns:
         A list of node types
         """
+
     def has_edge(self, src, dst):
         """
         Returns true if the graph contains the specified edge
@@ -3653,8 +3961,10 @@ class PersistentGraph:
         Returns:
          true if the graph contains the specified edge, false otherwise
         """
+
     def has_layer(self, name):
-        """Check if GraphView has the layer `"name"`"""
+        """ Check if GraphView has the layer `"name"`"""
+
     def has_node(self, id):
         """
         Returns true if the graph contains the specified node
@@ -3665,6 +3975,7 @@ class PersistentGraph:
         Returns:
           true if the graph contains the specified node, false otherwise
         """
+
     def import_edge(self, edge, force=False):
         """
         Import a single edge into the graph.
@@ -3680,6 +3991,7 @@ class PersistentGraph:
         Returns:
             Result<EdgeView<Graph, Graph>, GraphError> - A Result object which is Ok if the edge was successfully imported, and Err otherwise.
         """
+
     def import_edges(self, edges, force=False):
         """
         Import multiple edges into the graph.
@@ -3693,6 +4005,7 @@ class PersistentGraph:
             force (boolean) - An optional boolean flag indicating whether to force the import of the edges.
 
         """
+
     def import_node(self, node, force=False):
         """
         Import a single node into the graph.
@@ -3707,6 +4020,7 @@ class PersistentGraph:
         Returns:
             Result<NodeView<Graph, Graph>, GraphError> - A Result object which is Ok if the node was successfully imported, and Err otherwise.
         """
+
     def import_nodes(self, nodes, force=False):
         """
         Import multiple nodes into the graph.
@@ -3720,6 +4034,7 @@ class PersistentGraph:
             force (boolean) - An optional boolean flag indicating whether to force the import of the nodes.
 
         """
+
     def index(self):
         """
         Indexes all node and edge properties.
@@ -3729,6 +4044,7 @@ class PersistentGraph:
         Returns:
            GraphIndex - Returns a GraphIndex
         """
+
     @property
     def latest_date_time(self):
         """
@@ -3737,6 +4053,7 @@ class PersistentGraph:
         Returns:
             the datetime of the latest activity in the graph
         """
+
     @property
     def latest_time(self):
         """
@@ -3745,6 +4062,7 @@ class PersistentGraph:
         Returns:
             the timestamp of the latest activity in the graph
         """
+
     def layer(self, name):
         """
          Return a view of GraphView containing the layer `"name"`
@@ -3753,6 +4071,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     def layers(self, names):
         """
          Return a view of GraphView containing all layers `names`
@@ -3764,6 +4083,7 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
+
     @staticmethod
     def load_cached(path):
         """
@@ -3778,9 +4098,8 @@ class PersistentGraph:
         Returns:
            PersistentGraph
         """
-    def load_edge_deletions_from_pandas(
-        self, df, time, src, dst, layer=None, layer_col=None
-    ):
+
+    def load_edge_deletions_from_pandas(self, df, time, src, dst, layer=None, layer_col=None):
         """
         Load edges deletions from a Pandas DataFrame into the graph.
 
@@ -3797,9 +4116,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edge_deletions_from_parquet(
-        self, parquet_path, time, src, dst, layer=None, layer_col=None
-    ):
+
+    def load_edge_deletions_from_parquet(self, parquet_path, time, src, dst, layer=None, layer_col=None):
         """
         Load edges deletions from a Parquet file into the graph.
 
@@ -3816,16 +4134,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edge_props_from_pandas(
-        self,
-        df,
-        src,
-        dst,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edge_props_from_pandas(self, df, src, dst, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edge properties from a Pandas DataFrame.
 
@@ -3844,16 +4154,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edge_props_from_parquet(
-        self,
-        parquet_path,
-        src,
-        dst,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edge_props_from_parquet(self, parquet_path, src, dst, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edge properties from parquet file
 
@@ -3872,18 +4174,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edges_from_pandas(
-        self,
-        df,
-        time,
-        src,
-        dst,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edges_from_pandas(self, df, time, src, dst, properties=None, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edges from a Pandas DataFrame into the graph.
 
@@ -3903,18 +4195,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_edges_from_parquet(
-        self,
-        parquet_path,
-        time,
-        src,
-        dst,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-        layer=None,
-        layer_col=None,
-    ):
+
+    def load_edges_from_parquet(self, parquet_path, time, src, dst, properties=None, constant_properties=None, shared_constant_properties=None, layer=None, layer_col=None):
         """
         Load edges from a Parquet file into the graph.
 
@@ -3934,6 +4216,7 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
+
     @staticmethod
     def load_from_file(path):
         """
@@ -3945,15 +4228,8 @@ class PersistentGraph:
         Returns:
            PersistentGraph
         """
-    def load_node_props_from_pandas(
-        self,
-        df,
-        id,
-        node_type=None,
-        node_type_col=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_node_props_from_pandas(self, df, id, node_type=None, node_type_col=None, constant_properties=None, shared_constant_properties=None):
         """
         Load node properties from a Pandas DataFrame.
 
@@ -3971,15 +4247,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_node_props_from_parquet(
-        self,
-        parquet_path,
-        id,
-        node_type=None,
-        node_type_col=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_node_props_from_parquet(self, parquet_path, id, node_type=None, node_type_col=None, constant_properties=None, shared_constant_properties=None):
         """
         Load node properties from a parquet file.
 
@@ -3997,17 +4266,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_nodes_from_pandas(
-        self,
-        df,
-        time,
-        id,
-        node_type=None,
-        node_type_col=None,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_nodes_from_pandas(self, df, time, id, node_type=None, node_type_col=None, properties=None, constant_properties=None, shared_constant_properties=None):
         """
         Load nodes from a Pandas DataFrame into the graph.
 
@@ -4026,17 +4286,8 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
-    def load_nodes_from_parquet(
-        self,
-        parquet_path,
-        time,
-        id,
-        node_type=None,
-        node_type_col=None,
-        properties=None,
-        constant_properties=None,
-        shared_constant_properties=None,
-    ):
+
+    def load_nodes_from_parquet(self, parquet_path, time, id, node_type=None, node_type_col=None, properties=None, constant_properties=None, shared_constant_properties=None):
         """
         Load nodes from a Parquet file into the graph.
 
@@ -4055,6 +4306,7 @@ class PersistentGraph:
         Raises:
             GraphError: If the operation fails.
         """
+
     def materialize(self):
         """
         Returns a 'materialized' clone of the graph view - i.e. a new graph with a copy of the data seen within the view instead of just a mask over the original graph
@@ -4062,6 +4314,7 @@ class PersistentGraph:
         Returns:
            GraphView - Returns a graph clone
         """
+
     def node(self, id):
         """
         Gets the node with the specified id
@@ -4072,6 +4325,7 @@ class PersistentGraph:
         Returns:
           the node with the specified id, or None if the node does not exist
         """
+
     @property
     def nodes(self):
         """
@@ -4080,6 +4334,7 @@ class PersistentGraph:
         Returns:
          the nodes in the graph
         """
+
     @property
     def properties(self):
         """
@@ -4089,6 +4344,7 @@ class PersistentGraph:
         Returns:
            HashMap<String, Prop> - Properties paired with their names
         """
+
     def rolling(self, window, step=None):
         """
         Creates a `WindowSet` with the given `window` size and optional `step` using a rolling window.
@@ -4102,6 +4358,7 @@ class PersistentGraph:
         Returns:
             A `WindowSet` object.
         """
+
     def save_to_file(self, path):
         """
          Saves the PersistentGraph to the given path.
@@ -4109,6 +4366,7 @@ class PersistentGraph:
         Arguments:
          path (str): The path to the file.
         """
+
     def serialise(self):
         """
          Serialise PersistentGraph to bytes.
@@ -4116,6 +4374,7 @@ class PersistentGraph:
         Returns:
           Bytes
         """
+
     def shrink_end(self, end):
         """
         Set the end of the window to the smaller of `end` and `self.end()`
@@ -4125,6 +4384,7 @@ class PersistentGraph:
         Returns:
              A GraphView object.
         """
+
     def shrink_start(self, start):
         """
         Set the start of the window to the larger of `start` and `self.start()`
@@ -4135,6 +4395,7 @@ class PersistentGraph:
         Returns:
              A GraphView object.
         """
+
     def shrink_window(self, start, end):
         """
         Shrink both the start and end of the window (same as calling `shrink_start` followed by `shrink_end` but more efficient)
@@ -4142,6 +4403,7 @@ class PersistentGraph:
         Arguments:
 
         """
+
     @property
     def start(self):
         """
@@ -4150,6 +4412,7 @@ class PersistentGraph:
         Returns:
             The earliest time that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     @property
     def start_date_time(self):
         """
@@ -4158,6 +4421,7 @@ class PersistentGraph:
         Returns:
              The earliest datetime that this GraphView is valid or None if the GraphView is valid for all times.
         """
+
     def subgraph(self, nodes):
         """
         Returns a subgraph given a set of nodes
@@ -4168,6 +4432,7 @@ class PersistentGraph:
         Returns:
            GraphView - Returns the subgraph
         """
+
     def subgraph_node_types(self, node_types):
         """
         Returns a subgraph filtered by node types given a set of node types
@@ -4178,14 +4443,8 @@ class PersistentGraph:
         Returns:
            GraphView - Returns the subgraph
         """
-    def to_networkx(
-        self,
-        explode_edges=False,
-        include_node_properties=True,
-        include_edge_properties=True,
-        include_update_history=True,
-        include_property_history=True,
-    ):
+
+    def to_networkx(self, explode_edges=False, include_node_properties=True, include_edge_properties=True, include_update_history=True, include_property_history=True):
         """
         Returns a graph with NetworkX.
 
@@ -4203,18 +4462,8 @@ class PersistentGraph:
             Returns:
                 A Networkx MultiDiGraph.
         """
-    def to_pyvis(
-        self,
-        explode_edges=False,
-        edge_color="#000000",
-        shape=None,
-        node_image=None,
-        edge_weight=None,
-        edge_label=None,
-        colour_nodes_by_type=False,
-        notebook=False,
-        **kwargs,
-    ):
+
+    def to_pyvis(self, explode_edges=False, edge_color="#000000", shape=None, node_image=None, edge_weight=None, edge_label=None, colour_nodes_by_type=False, notebook=False, **kwargs):
         """
         Draw a graph with PyVis.
         Pyvis is a required dependency. If you intend to use this function make sure that you install Pyvis
@@ -4238,9 +4487,11 @@ class PersistentGraph:
             Returns:
                 A pyvis network
         """
+
     @property
     def unique_layers(self):
         """Return all the layer ids in the graph"""
+
     def update_constant_properties(self, properties):
         """
         Updates static properties to the graph.
@@ -4251,6 +4502,7 @@ class PersistentGraph:
         Returns:
            None
         """
+
     def valid_layers(self, names):
         """
          Return a view of GraphView containing all layers `names`
@@ -4262,16 +4514,8 @@ class PersistentGraph:
         Returns:
              GraphView: The layered view
         """
-    def vectorise(
-        self,
-        embedding,
-        cache=None,
-        overwrite_cache=False,
-        graph_document=None,
-        node_document=None,
-        edge_document=None,
-        verbose=False,
-    ):
+
+    def vectorise(self, embedding, cache=None, overwrite_cache=False, graph_document=None, node_document=None, edge_document=None, verbose=False):
         """
         Create a VectorisedGraph from the current graph
 
@@ -4286,6 +4530,7 @@ class PersistentGraph:
         Returns:
           A VectorisedGraph with all the documents/embeddings computed and with an initial empty selection
         """
+
     def window(self, start, end):
         """
          Create a view of the GraphView including all events between `start` (inclusive) and `end` (exclusive)
@@ -4297,9 +4542,11 @@ class PersistentGraph:
         Returns:
         r    A GraphView object.
         """
+
     @property
     def window_size(self):
-        """Get the window size (difference between start and end) for this GraphView"""
+        """ Get the window size (difference between start and end) for this GraphView"""
+
     def write_updates(self):
         """Persist the new updates by appending them to the cache file."""
 
@@ -4308,11 +4555,14 @@ class Properties:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def as_dict(self):
         """Convert properties view to a dict"""
+
     @property
     def constant(self):
         """Get a view of the constant properties (meta-data) only."""
+
     def get(self, key):
         """
         Get property value.
@@ -4320,13 +4570,17 @@ class Properties:
         First searches temporal properties and returns latest value if it exists.
         If not, it falls back to static properties.
         """
+
     def items(self):
         """Get a list of key-value pairs"""
+
     def keys(self):
         """Get the names for all properties (includes temporal and static properties)"""
+
     @property
     def temporal(self):
         """Get a view of the temporal properties only."""
+
     def values(self):
         """
         Get the values of the properties
@@ -4340,9 +4594,12 @@ class PyDirection:
 
     def __init__(self, direction):
         """Initialize self.  See help(type(self)) for accurate signature."""
-    def as_str(self): ...
+
+    def as_str(self):
+        ...
 
 class PyGraphEncoder:
+
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
 
@@ -4351,8 +4608,10 @@ class TemporalProp:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def at(self, t):
         """Get the value of the property at time `t`"""
+
     def average(self):
         """
         Compute the average of all property values. Alias for mean().
@@ -4360,6 +4619,7 @@ class TemporalProp:
         Returns:
             Prop: The average of each property values, or None if count is zero.
         """
+
     def count(self):
         """
         Count the number of properties.
@@ -4367,14 +4627,19 @@ class TemporalProp:
         Returns:
             int: The number of properties.
         """
+
     def history(self):
         """Get the timestamps at which the property was updated"""
+
     def history_date_time(self):
         """Get the timestamps at which the property was updated"""
+
     def items(self):
         """List update timestamps and corresponding property values"""
+
     def items_date_time(self):
         """List update timestamps and corresponding property values"""
+
     def max(self):
         """
         Find the maximum property value and its associated time.
@@ -4382,6 +4647,7 @@ class TemporalProp:
         Returns:
             (i64, Prop): A tuple containing the time and the maximum property value.
         """
+
     def mean(self):
         """
         Compute the mean of all property values. Alias for mean().
@@ -4389,6 +4655,7 @@ class TemporalProp:
         Returns:
             Prop: The mean of each property values, or None if count is zero.
         """
+
     def median(self):
         """
         Compute the median of all property values.
@@ -4396,6 +4663,7 @@ class TemporalProp:
         Returns:
             (i64, Prop): A tuple containing the time and the median property value, or None if empty
         """
+
     def min(self):
         """
         Find the minimum property value and its associated time.
@@ -4403,7 +4671,10 @@ class TemporalProp:
         Returns:
             (i64, Prop): A tuple containing the time and the minimum property value.
         """
-    def ordered_dedupe(self, latest_time): ...
+
+    def ordered_dedupe(self, latest_time):
+        ...
+
     def sum(self):
         """
         Compute the sum of all property values.
@@ -4411,9 +4682,13 @@ class TemporalProp:
         Returns:
             Prop: The sum of all property values.
         """
-    def unique(self): ...
+
+    def unique(self):
+        ...
+
     def value(self):
         """Get the latest value of the property"""
+
     def values(self):
         """Get the property values for each update"""
 
@@ -4422,6 +4697,7 @@ class TemporalProperties:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def get(self, key):
         """
         get(key: str) -> Optional[TemporalProp]
@@ -4431,6 +4707,7 @@ class TemporalProperties:
         Returns:
             the property view if it exists, otherwise `None`
         """
+
     def histories(self):
         """
         Get the histories of all properties
@@ -4438,6 +4715,7 @@ class TemporalProperties:
         Returns:
             dict[str, list[(int, Any)]]: the mapping of property keys to histories
         """
+
     def histories_date_time(self):
         """
         Get the histories of all properties
@@ -4445,10 +4723,13 @@ class TemporalProperties:
         Returns:
             dict[str, list[(datetime, Any)]]: the mapping of property keys to histories
         """
+
     def items(self):
         """List the property keys together with the corresponding values"""
+
     def keys(self):
         """List the available property keys"""
+
     def latest(self):
         """
         Get the latest value of all properties
@@ -4456,6 +4737,7 @@ class TemporalProperties:
         Returns:
             dict[str, Any]: the mapping of property keys to latest values
         """
+
     def values(self):
         """
         List the values of the properties
