+++ python/python/raphtory/algorithms/__init__.pyi
@@ -68,12 +68,12 @@ def betweenness_centrality(g, k=None, normalized=True):
         AlgorithmResult[float]: Returns an `AlgorithmResult` containing the betweenness centrality of each node.
     """
 
-def cohesive_fruchterman_reingold(
-    graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1
-):
+def cohesive_fruchterman_reingold(graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1):
     """Cohesive version of `fruchterman_reingold` that adds virtual edges between isolated nodes"""
 
-def connected_components(g): ...
+def connected_components(g):
+    ...
+
 def degree_centrality(g, threads=None):
     """
     Computes the degree centrality of all nodes in the graph. The values are normalized
@@ -88,9 +88,7 @@ def degree_centrality(g, threads=None):
         AlgorithmResult<String, OrderedFloat<f64>>: A result containing a mapping of node names to the computed sum of their associated degree centrality.
     """
 
-def dijkstra_single_source_shortest_paths(
-    g, source, targets, direction=..., weight=...
-):
+def dijkstra_single_source_shortest_paths(g, source, targets, direction=..., weight=...):
     """
     Finds the shortest paths from a single source to multiple targets in a graph.
 
@@ -120,9 +118,7 @@ def directed_graph_density(g):
         float : Directed graph density of G.
     """
 
-def fruchterman_reingold(
-    graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1
-):
+def fruchterman_reingold(graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1):
     """
     Fruchterman Reingold layout algorithm
 
@@ -453,15 +449,7 @@ def strongly_connected_components(g):
         Vec<Vec<u64>> : List of strongly connected nodes identified by ids
     """
 
-def temporal_SEIR(
-    graph,
-    seeds,
-    infection_prob,
-    initial_infection,
-    recovery_rate=None,
-    incubation_rate=None,
-    rng_seed=None,
-):
+def temporal_SEIR(graph, seeds, infection_prob, initial_infection, recovery_rate=None, incubation_rate=None, rng_seed=None):
     """
     Simulate an SEIR dynamic on the network
 
@@ -485,7 +473,7 @@ def temporal_SEIR(
 
     Returns:
         AlgorithmResult[Infected]: Returns an `Infected` object for each infected node with attributes
-
+    
         `infected`: the time stamp of the infection event
 
         `active`: the time stamp at which the node actively starts spreading the infection (i.e., the end of the incubation period)
