+++ python/python/raphtory/graph_loader/__init__.pyi
@@ -54,7 +54,9 @@ def lotr_graph():
       A Graph containing the LOTR dataset
     """
 
-def neo4j_movie_graph(uri, username, password, database=...): ...
+def neo4j_movie_graph(uri, username, password, database=...):
+    ...
+
 def reddit_hyperlink_graph(timeout_seconds=600):
     """
     Load (a subset of) Reddit hyperlinks dataset into a graph.
@@ -94,5 +96,8 @@ def reddit_hyperlink_graph(timeout_seconds=600):
      A Graph containing the Reddit hyperlinks dataset
     """
 
-def reddit_hyperlink_graph_local(file_path): ...
-def stable_coin_graph(path=None, subset=None): ...
+def reddit_hyperlink_graph_local(file_path):
+    ...
+
+def stable_coin_graph(path=None, subset=None):
+    ...
