+++ python/python/raphtory/graphql/__init__.pyi
@@ -10,15 +10,9 @@
 class GraphServer:
     """A class for defining and running a Raphtory GraphQL server"""
 
-    def __init__(
-        self,
-        work_dir,
-        cache_capacity=None,
-        cache_tti_seconds=None,
-        log_level=None,
-        config_path=None,
-    ):
+    def __init__(self, work_dir, cache_capacity=None, cache_tti_seconds=None, log_level=None, config_path=None):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def run(self, port=1736, timeout_ms=...):
         """
         Run the server until completion.
@@ -26,6 +20,7 @@ class GraphServer:
         Arguments:
           * `port`: the port to use (defaults to 1736).
         """
+
     def start(self, port=1736, timeout_ms=None):
         """
         Start the server and return a handle to it.
@@ -34,6 +29,7 @@ class GraphServer:
           * `port`: the port to use (defaults to 1736).
           * `timeout_ms`: wait for server to be online (defaults to 5000). The server is stopped if not online within timeout_ms but manages to come online as soon as timeout_ms finishes!
         """
+
     def with_document_search_function(self, name, input, function):
         """
         Register a function in the GraphQL schema for document search over a graph.
@@ -51,6 +47,7 @@ class GraphServer:
         Returns:
            A new server object containing the vectorised graphs.
         """
+
     def with_global_search_function(self, name, input, function):
         """
         Register a function in the GraphQL schema for document search among all the graphs.
@@ -68,15 +65,8 @@ class GraphServer:
         Returns:
            A new server object containing the vectorised graphs.
         """
-    def with_vectorised(
-        self,
-        cache,
-        graph_names=None,
-        embedding=None,
-        graph_document=None,
-        node_document=None,
-        edge_document=None,
-    ):
+
+    def with_vectorised(self, cache, graph_names=None, embedding=None, graph_document=None, node_document=None, edge_document=None):
         """
         Vectorise a subset of the graphs of the server.
 
@@ -104,8 +94,10 @@ class GraphqlGraphs:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def get(self, name):
         """Return the `VectorisedGraph` with name `name` or `None` if it doesn't exist"""
+
     def search_graph_documents(self, query, limit, window):
         """
         Return the top documents with the smallest cosine distance to `query`
@@ -118,6 +110,7 @@ class GraphqlGraphs:
         # Returns
           A list of documents
         """
+
     def search_graph_documents_with_scores(self, query, limit, window):
         """Same as `search_graph_documents` but it also returns the scores alongside the documents"""
 
@@ -126,6 +119,7 @@ class RaphtoryClient:
 
     def __init__(self, url):
         """Initialize self.  See help(type(self)) for accurate signature."""
+
     def copy_graph(self, path, new_path):
         """
         Copy graph from a path `path` on the server to a `new_path` on the server
@@ -137,6 +131,7 @@ class RaphtoryClient:
         Returns:
            Copy status as boolean
         """
+
     def delete_graph(self, path):
         """
         Delete graph from a path `path` on the server
@@ -147,6 +142,7 @@ class RaphtoryClient:
         Returns:
            Delete status as boolean
         """
+
     def is_server_online(self):
         """
         Check if the server is online.
@@ -154,6 +150,7 @@ class RaphtoryClient:
         Returns:
            Returns true if server is online otherwise false.
         """
+
     def move_graph(self, path, new_path):
         """
         Move graph from a path `path` on the server to a `new_path` on the server
@@ -165,6 +162,20 @@ class RaphtoryClient:
         Returns:
            Move status as boolean
         """
+
+    def new_graph(self, path, graph_type):
+        """
+        Create a new Graph on the server at `path`
+
+        Arguments:
+          * `path`: the path of the graph to be created
+          * `graph_type`: the type of graph that should be created - this can be EVENT or PERSISTENT
+
+        Returns:
+           None
+
+        """
+
     def query(self, query, variables=None):
         """
         Make a graphQL query against the server.
@@ -176,6 +187,7 @@ class RaphtoryClient:
         Returns:
            The `data` field from the graphQL response.
         """
+
     def receive_graph(self, path):
         """
         Receive graph from a path `path` on the server
@@ -186,6 +198,7 @@ class RaphtoryClient:
         Returns:
            Graph as string
         """
+
     def send_graph(self, path, graph, overwrite=False):
         """
         Send a graph to the server
@@ -198,6 +211,7 @@ class RaphtoryClient:
         Returns:
            The `data` field from the graphQL response after executing the mutation.
         """
+
     def upload_graph(self, path, file_path, overwrite=False):
         """
         Upload graph file from a path `file_path` on the client
@@ -216,6 +230,9 @@ class RunningGraphServer:
 
     def __init__(self):
         """Initialize self.  See help(type(self)) for accurate signature."""
-    def get_client(self): ...
+
+    def get_client(self):
+        ...
+
     def stop(self):
         """Stop the server and wait for it to finish"""
