+++ raphtory-graphql/src/python/client/mod.rs
@@ -1,0 +1,1 @@
+pub mod raphtory_client;
