+++ raphtory-graphql/src/python/pymodule.rs
@@ -1,5 +1,7 @@
-use crate::python::graphql::{
-    PyGlobalPlugins, PyGraphServer, PyRaphtoryClient, PyRunningGraphServer,
+use crate::python::{
+    client::raphtory_client::PyRaphtoryClient,
+    global_plugins::PyGlobalPlugins,
+    server::{running_server::PyRunningGraphServer, server::PyGraphServer},
 };
 use pyo3::{prelude::PyModule, PyErr, Python};
 
@@ -9,5 +11,5 @@ pub fn base_graphql_module(py: Python<'_>) -> Result<&PyModule, PyErr> {
     graphql_module.add_class::<PyGraphServer>()?;
     graphql_module.add_class::<PyRunningGraphServer>()?;
     graphql_module.add_class::<PyRaphtoryClient>()?;
-    return Ok(graphql_module);
+    Ok(graphql_module)
 }
